move StandardizedVLLM to dedicated module with lazy vllm import
- Separated StandardizedVLLM into standardized_vllm.py to avoid importing
  vllm when not needed
- Moved load_model() to __init__.py with TYPE_CHECKING guard for vllm
- Added hf_kwargs_to_vllm_kwargs utility for argument translation
- Updated pyproject.toml: requires-python>=3.10, vllm uses nnsight[vllm]
- Updated README test instructions to use uv run pytest
